requirements.txt
streamlit>=1.28.0
tensorflow>=2.14.0
numpy>=1.24.0
pillow>=10.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
openai>=1.3.0
matplotlib>=3.8.0
imagehash>=4.3.0
tqdm>=4.66.0
seaborn>=0.12.0
scikit-learn>=1.3.0
plotly>=5.17.0
